Rewrote reviews/templates/reviews/search_results.html.
Now it's extends the base.html, if the search field of the form is empty the title of the page Book Search
otherwise Search results for smth. Included form tag, enable csrf protection, get the form from render template
added button. If the form isn't empty below the button a text will apperead.